fix: logout of unauthorized users
